Isolate boost::multi hdf template specialisations."
