+++ src/AFQMC/Drivers/AFQMCDriver.h
@@ -1,6 +1,7 @@
 #ifndef QMCPLUSPLUS_AFQMC_AFQMCDRIVER_H
 #define QMCPLUSPLUS_AFQMC_AFQMCDRIVER_H
 
+#include "io/hdf_multi.h"
 #include "io/hdf_archive.h"
 #include "mpi3/communicator.hpp"
