+++ src/AFQMC/Drivers/BenchmarkDriver.h
@@ -2,6 +2,7 @@
 #define QMCPLUSPLUS_AFQMC_BENCHMARKDRIVER_H
 
 #include<Message/MPIObjectBase.h>
+#include "io/hdf_multi.h"
 #include "io/hdf_archive.h"
 
 #include "AFQMC/config.h"
