+++ src/AFQMC/Estimators/EstimatorBase.h
@@ -6,6 +6,7 @@
 #include<iostream>
 #include<fstream>
 
+#include "io/hdf_multi.h"
 #include "io/hdf_archive.h"
 #include "OhmmsData/libxmldefs.h"
