+++ src/AFQMC/Estimators/OneRdmEstimator.h
@@ -8,6 +8,7 @@
 #include <iostream>
 #include <fstream>
 
+#include "io/hdf_multi.h"
 #include "io/hdf_archive.h"
 #include "OhmmsData/libxmldefs.h"
 #include "Utilities/NewTimer.h"
