+++ src/AFQMC/Estimators/WalkerDMEstimator.h
@@ -9,6 +9,7 @@
 #include <iostream>
 #include <fstream>
 
+#include "io/hdf_multi.h"
 #include "io/hdf_archive.h"
 #include "OhmmsData/libxmldefs.h"
 #include "Utilities/NewTimer.h"
