+++ src/AFQMC/HamiltonianOperations/HamOpsIO.hpp
@@ -18,6 +18,7 @@
 
 #include<fstream>
 
+#include "io/hdf_multi.h"
 #include "io/hdf_archive.h"
 
 #include "AFQMC/config.h"
