+++ src/AFQMC/Walkers/WalkerIO.hpp
@@ -20,6 +20,7 @@
 #include<cstdlib>
 #include<vector>
 #include<type_traits>
+#include "type_traits/container_proxy_multi.h"
 
 #include "Configuration.h"
 #include "AFQMC/config.h"
