+++ src/type_traits/container_proxy.h
@@ -24,8 +24,6 @@
 #if defined(HAVE_LIBBOOST)
 #include <boost/multi_array.hpp>
 #endif
-#include "multi/array.hpp"
-#include "multi/array_ref.hpp"
 
 namespace qmcplusplus
 {
@@ -292,65 +290,6 @@ struct container_proxy<boost::multi_array_ref<T,2> >
 };
 #endif
 
-template<typename T, class Alloc>
-struct container_proxy<boost::multi::array<T,2,Alloc> >
-{
-  enum {DIM=scalar_traits<T>::DIM};
-  typedef typename container_proxy<T>::pointer pointer;
-  boost::multi::array<T,2,Alloc>& ref;
-  inline container_proxy(boost::multi::array<T,2,Alloc>& a):ref(a) {}
-  inline size_t size() const
-  {
-    return ref.num_elements()*DIM;
-  }
-  inline pointer data()
-  {
-    //using detail::to_address;
-    //return scalar_traits<T>::get_address(to_address(ref.origin()));
-    return scalar_traits<T>::get_address(std::addressof(*ref.origin()));
-  }
-  inline void resize(size_t n)
-  {
-    APP_ABORT(" Error: Can not resize container_proxy<boost::multi::array<T,D,Alloc> >. \n");
-  }
-  template<typename I>
-  inline void resize(I* n, int d)
-  {
-    if(d < 2)
-      APP_ABORT(" Error: Inconsistent dimension in container_proxy<boost::multi::array<T,D,Alloc> >::resize(I*,int). \n");
-    ref.reextent({n[0],n[1]});
-  }
-};
-
-template<typename T>
-struct container_proxy<boost::multi::array_ref<T,2> >
-{
-  enum {DIM=scalar_traits<T>::DIM};
-  typedef typename container_proxy<T>::pointer pointer;
-  boost::multi::array_ref<T,2>& ref;
-  inline container_proxy(boost::multi::array_ref<T,2>& a):ref(a) {}
-  inline size_t size() const
-  {
-    return ref.num_elements()*DIM;
-  }
-  inline pointer data()
-  {
-    //using detail::to_address;
-    //return scalar_traits<T>::get_address(to_address(ref.origin()));
-    return scalar_traits<T>::get_address(std::addressof(*ref.origin()));
-  }
-  inline void resize(size_t n)
-  {
-    APP_ABORT(" Error: Can not resize container_proxy<boost::multi::array_ref<T,D> >. \n");
-  }
-
-  template<typename I>
-  inline void resize(I* n, int d)
-  {
-    APP_ABORT(" Error: Can not resize container_proxy<boost::multi::array_ref<T,D> >. \n");
-  }
-
-};
 
 }
 #endif
